refactor: merge getPageSchema + pageConfig types
