+++ app/domains/beratungshilfe/vorabcheck/userData.ts
@@ -1,4 +1,4 @@
-import type { UserDataFromPagesSchema } from "~/domains/pageConfig";
+import type { UserDataFromPagesSchema } from "~/domains/pageSchemas";
 import type { beratungshilfeVorabcheckPages } from "./pages";
 
 export type BeratungshilfeVorabcheckUserData = UserDataFromPagesSchema<
