+++ app/domains/pageSchemas.ts
@@ -1,4 +1,20 @@
 import type { z } from "zod";
+import { beratungshilfeVorabcheckPages } from "./beratungshilfe/vorabcheck/pages";
+import { flowIdFromPathname, parsePathname, type FlowId } from "./flowIds";
+
+const pages: Partial<Record<FlowId, PagesConfig>> = {
+  "/beratungshilfe/vorabcheck": beratungshilfeVorabcheckPages,
+} as const;
+
+export function getPageSchema(pathname: string) {
+  const flowId = flowIdFromPathname(pathname);
+  if (!flowId || !(flowId in pages)) return undefined;
+  const { stepId } = parsePathname(pathname);
+  const stepIdWithoutLeadingSlash = stepId.slice(1);
+  return Object.values(pages[flowId] ?? {}).find(
+    (page) => page.url === stepIdWithoutLeadingSlash,
+  )?.pageSchema;
+}
 
 // TODO: better specify PageSchema to specify enums, strings, ...
 export type PageSchema = Record<string, z.ZodTypeAny>;
